Install Conda requirements.

Install Conda requirements

install_CondaTools(
  tools,
  env,
  channels = NULL,
  pathToMiniConda = NULL,
  updateEnv = FALSE,
  search = TRUE
)

Arguments

tools
Vector of software to install using conda.

env
Name of Conda environment to install tools into.

channels
Additional channels for miniconda (bioconda defaults and conda-forge are included automatically)

pathToMiniConda
NULL Path to miniconda installation

updateEnv
Update existing package's conda environment if already installed.

search
Whether to search for the package name and version before installing. It is highly recommended this be set to TRUE as information about available versions or similar packages will be included in the output if the exact match is not found.

Value

Nothing returned. Output written to file.

Author

Thomas Carroll

Examples

condaDir <- file.path(tempdir(), "r-miniconda")
condaPaths <- install_CondaTools("salmon", "salmon", pathToMiniConda = condaDir)

#> The environment salmon does not currently exist and will be created. 

#> The package(s) (salmon) are in the salmon environment. 

#> [1] "Conda and Environment Information"
#> $pathToConda
#> [1] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpyPYTTX/r-miniconda/bin/conda"
#> 
#> $environment
#> [1] "salmon"
#> 
#> $pathToEnvBin
#> [1] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpyPYTTX/r-miniconda/envs/salmon/bin"
#> 

system2(file.path(condaPaths$pathToEnvBin, "salmon"), args = "--help")